Handle migration in the interface
