+++ internal/db/db.go
@@ -4,15 +4,18 @@ import (
 	"fmt"
 
 	"github.com/redhatinsights/platform-changelog-go/internal/config"
+	"github.com/redhatinsights/platform-changelog-go/internal/logging"
 	l "github.com/redhatinsights/platform-changelog-go/internal/logging"
-
+	"github.com/redhatinsights/platform-changelog-go/internal/models"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 )
 
-var DB *gorm.DB
+type DBConnectorImpl struct {
+	db *gorm.DB
+}
 
-func DbConnect(cfg *config.Config) {
+func NewDBConnector(cfg *config.Config) DBConnector {
 	var (
 		user     = cfg.DatabaseConfig.DBUser
 		password = cfg.DatabaseConfig.DBPassword
@@ -27,7 +30,18 @@ func DbConnect(cfg *config.Config) {
 		l.Log.Fatal(err)
 	}
 
-	DB = db
-
 	l.Log.Info("DB initialization complete")
+
+	return &DBConnectorImpl{db: db}
+}
+
+func (conn *DBConnectorImpl) Migrate() {
+	conn.db.Exec("CREATE TYPE timeline_type AS ENUM ('unknown', 'commit', 'deploy')")
+
+	conn.db.AutoMigrate(
+		&models.Services{},
+		&models.Timelines{},
+	)
+
+	logging.Log.Info("DB Migration Complete")
 }
