+++ internal/migration/main.go
@@ -4,36 +4,29 @@ import (
 	"github.com/redhatinsights/platform-changelog-go/internal/config"
 	"github.com/redhatinsights/platform-changelog-go/internal/db"
 	"github.com/redhatinsights/platform-changelog-go/internal/logging"
-	"github.com/redhatinsights/platform-changelog-go/internal/models"
-
-	"gorm.io/gorm"
 )
+
+var conn db.DBConnector
 
 func main() {
 	logging.InitLogger()
 
 	cfg := config.Get()
 
-	db.DbConnect(cfg)
+	conn = db.NewDBConnector(cfg)
 
-	// Set up TimelineType Enum (gorm doesn't have a function for this)
-	db.DB.Exec("CREATE TYPE timeline_type AS ENUM ('unknown', 'commit', 'deploy')")
-
-	db.DB.AutoMigrate(
-		&models.Services{},
-		&models.Timelines{},
-	)
+	conn.Migrate()
 
 	logging.Log.Info("DB Migration Complete")
 
-	reconcileServices(db.DB, cfg)
+	reconcileServices(cfg)
 }
 
-func reconcileServices(g *gorm.DB, cfg *config.Config) {
+func reconcileServices(cfg *config.Config) {
 	for key, service := range cfg.Services {
-		_, rowsAffected, _ := db.GetServiceByName(g, key)
+		_, rowsAffected, _ := conn.GetServiceByName(key)
 		if rowsAffected == 0 {
-			_, service := db.CreateServiceTableEntry(g, key, service)
+			_, service := conn.CreateServiceTableEntry(key, service)
 			logging.Log.Info("Created service: ", service)
 		} else {
 			logging.Log.Info("Service already exists: ", service.DisplayName)
